Move useSettingsValueWithSetter to useSettings
